Refactor Chunk2Doc from base to annotator
